@@ -1,6 +1,6 @@
 //File generated by vaweei CLI
 import { Component, OnInit, Output, EventEmitter, ViewChild, ElementRef, HostListener, OnDestroy, Input, AfterViewInit } from '@angular/core';
-import { FormGroup, FormControl, Validators, FormBuilder } from '@angular/forms';
+import { UntypedFormGroup, UntypedFormControl, Validators, UntypedFormBuilder } from '@angular/forms';
 import { ActivatedRoute, Router } from '@angular/router';
 import { Subscription } from 'rxjs';
 import { filter } from 'rxjs/operators';
@@ -19,7 +19,7 @@ import { UtilService } from 'src/app/services/util.service';
 export class Hobbies_InterestFormViewComponent implements OnInit, OnDestroy, AfterViewInit {
 
   @ViewChild('firstInput', { static: false }) firstInput!: ElementRef;
-  formGroup: FormGroup;
+  formGroup: UntypedFormGroup;
   @Input() modalConfigParent: any;
   @Input() nameForm: String = '';
 
@@ -30,13 +30,13 @@ export class Hobbies_InterestFormViewComponent implements OnInit, OnDestroy, Aft
     private router : Router,
     private route: ActivatedRoute,
     private headerService : HeaderService,
-    private fb: FormBuilder,
+    private fb: UntypedFormBuilder,
     private utilService: UtilService
   ) {
     this.setFocus();
     this.formGroup = this.fb.group({
-      code: new FormControl(null, [Validators.required, Validators.maxLength(250)]),
-      name: new FormControl(null, [Validators.required, Validators.maxLength(250)]),
+      code: new UntypedFormControl(null, [Validators.required, Validators.maxLength(250)]),
+      name: new UntypedFormControl(null, [Validators.required, Validators.maxLength(250)]),
     });
   }
 
