@@ -1,5 +1,5 @@
 import { Component, OnInit, Output, EventEmitter, ViewChild, ElementRef, HostListener, OnDestroy, Input, AfterViewInit } from '@angular/core';
-import { FormGroup, FormControl, Validators, FormBuilder } from '@angular/forms';
+import { UntypedFormGroup, UntypedFormControl, Validators, UntypedFormBuilder } from '@angular/forms';
 import { ActivatedRoute } from '@angular/router'; 
 import { CATALOGS } from 'src/app/utils/setup/routes.enum';
 
@@ -18,7 +18,7 @@ import { filter } from 'rxjs/operators';
 export class HobbiesInterestsFormViewComponent implements OnInit, OnDestroy, AfterViewInit {
 
   @ViewChild('firstInput', { static: false }) firstInput!: ElementRef;
-  formGroup: FormGroup;
+  formGroup: UntypedFormGroup;
   @Input() modalConfigParent: any;
   @Input() nameForm: String = '';
   loading = false;
@@ -46,12 +46,12 @@ export class HobbiesInterestsFormViewComponent implements OnInit, OnDestroy, Aft
     // private router: Router,
     // private sessionService: SessionService,
     // private addressService: AddressesService,
-    private fb: FormBuilder,
+    private fb: UntypedFormBuilder,
     // private stepperFisherProducerForm: StepperFisherProducerFormService
   ) {
     this.formGroup = this.fb.group({
-      hobbies_interest: new FormControl(null, [Validators.required, Validators.maxLength(250)]),
-      notes: new FormControl(null, [])
+      hobbies_interest: new UntypedFormControl(null, [Validators.required, Validators.maxLength(250)]),
+      notes: new UntypedFormControl(null, [])
     });
   }
   ngAfterViewInit(): void {
