@@ -1,5 +1,5 @@
 import { Component, OnInit, Output, EventEmitter, ViewChild, ElementRef, HostListener, OnDestroy, Input, AfterViewInit } from '@angular/core';
-import { FormGroup, FormControl, Validators, FormBuilder } from '@angular/forms';
+import { UntypedFormGroup, UntypedFormControl, Validators, UntypedFormBuilder } from '@angular/forms';
 import { ActivatedRoute } from '@angular/router'; 
 
 //SERVICES
@@ -17,7 +17,7 @@ import { filter } from 'rxjs/operators';
 export class NonverbalLanguageFormViewComponent implements OnInit, OnDestroy, AfterViewInit {
 
   @ViewChild('firstInput', { static: false }) firstInput!: ElementRef;
-  formGroup: FormGroup;
+  formGroup: UntypedFormGroup;
   @Input() modalConfigParent: any;
   @Input() nameForm: String = '';
   loading = false;
@@ -28,14 +28,14 @@ export class NonverbalLanguageFormViewComponent implements OnInit, OnDestroy, Af
     private patientService: PatientService,
     private route: ActivatedRoute,
     private utilService: UtilService,
-    private fb: FormBuilder,
+    private fb: UntypedFormBuilder,
     // private stepperFisherProducerForm: StepperFisherProducerFormService
   ) {
     this.formGroup = this.fb.group({
-      time: new FormControl(false, []),
-      person: new FormControl(false, []),
-      space: new FormControl(false, []),
-      notes: new FormControl(null, [])
+      time: new UntypedFormControl(false, []),
+      person: new UntypedFormControl(false, []),
+      space: new UntypedFormControl(false, []),
+      notes: new UntypedFormControl(null, [])
     });
   }
   
