@@ -1,5 +1,5 @@
 import { Component, OnInit, Output, EventEmitter, ViewChild, ElementRef, HostListener, OnDestroy, Input, AfterViewInit } from '@angular/core';
-import { FormGroup, FormControl, Validators, FormBuilder } from '@angular/forms';
+import { UntypedFormGroup, UntypedFormControl, Validators, UntypedFormBuilder } from '@angular/forms';
 import { ActivatedRoute, Router } from '@angular/router'; 
 import { CATALOGS } from 'src/app/utils/setup/routes.enum';
 
@@ -20,7 +20,7 @@ import { filter } from 'rxjs/operators';
 export class TemplateConfigurationFormViewComponent implements OnInit, OnDestroy, AfterViewInit {
 
   @ViewChild('firstInput', { static: false }) firstInput!: ElementRef;
-  formGroup: FormGroup;
+  formGroup: UntypedFormGroup;
   @Input() modalConfigParent: any;
   @Input() nameForm: String = '';
   loading = false;
@@ -37,13 +37,13 @@ export class TemplateConfigurationFormViewComponent implements OnInit, OnDestroy
     private route: ActivatedRoute,
     private headerService : HeaderService,
     private utilService: UtilService,
-    private fb: FormBuilder
+    private fb: UntypedFormBuilder
   ) {
     this.formGroup = this.fb.group({
-      name: new FormControl(null, [Validators.required, Validators.maxLength(250)]),
-      description: new FormControl(null, [Validators.required, Validators.maxLength(250)]),
-      code: new FormControl(null, [Validators.required, Validators.maxLength(250)]),
-      emotions: new FormControl(null, [Validators.required, Validators.maxLength(250)])
+      name: new UntypedFormControl(null, [Validators.required, Validators.maxLength(250)]),
+      description: new UntypedFormControl(null, [Validators.required, Validators.maxLength(250)]),
+      code: new UntypedFormControl(null, [Validators.required, Validators.maxLength(250)]),
+      emotions: new UntypedFormControl(null, [Validators.required, Validators.maxLength(250)])
     });
   }
 
