@@ -1,6 +1,6 @@
 //File generated by vaweei CLI
 import { Component, OnInit, Output, EventEmitter, ViewChild, ElementRef, HostListener, OnDestroy, Input, AfterViewInit } from '@angular/core';
-import { FormGroup, FormControl, Validators, FormBuilder } from '@angular/forms';
+import { UntypedFormGroup, UntypedFormControl, Validators, UntypedFormBuilder } from '@angular/forms';
 import { ActivatedRoute, Router } from '@angular/router';
 import { Subscription } from 'rxjs';
 import { filter } from 'rxjs/operators';
@@ -19,7 +19,7 @@ import { UtilService } from 'src/app/services/util.service';
 export class TrackingFormViewComponent implements OnInit, OnDestroy, AfterViewInit {
 
   @ViewChild('firstInput', { static: false }) firstInput!: ElementRef;
-  formGroup: FormGroup;
+  formGroup: UntypedFormGroup;
   @Input() modalConfigParent: any;
   @Input() nameForm: String = '';
 
@@ -32,18 +32,18 @@ export class TrackingFormViewComponent implements OnInit, OnDestroy, AfterViewIn
     private router : Router,
     private route: ActivatedRoute,
     private headerService : HeaderService,
-    private fb: FormBuilder,
+    private fb: UntypedFormBuilder,
     private utilService: UtilService
   ) {
     this.setFocus();
     this.formGroup = this.fb.group({
-      patient:  new FormControl(null, [Validators.required, Validators.maxLength(250)]),
-      session_approach:  new FormControl(null, [Validators.required, Validators.maxLength(250)]),
-      clinical_progress:  new FormControl(null, [Validators.required, Validators.maxLength(250)]),
-      session_objective:  new FormControl(null, [Validators.required, Validators.maxLength(250)]),
-      state_of_mind:  new FormControl(null, [Validators.required, Validators.maxLength(250)]),
-      conducts_and_non_verbal_languages:  new FormControl(null, [Validators.required, Validators.maxLength(250)]),
-      observations:  new FormControl(null, [Validators.required, Validators.maxLength(250)]),
+      patient:  new UntypedFormControl(null, [Validators.required, Validators.maxLength(250)]),
+      session_approach:  new UntypedFormControl(null, [Validators.required, Validators.maxLength(250)]),
+      clinical_progress:  new UntypedFormControl(null, [Validators.required, Validators.maxLength(250)]),
+      session_objective:  new UntypedFormControl(null, [Validators.required, Validators.maxLength(250)]),
+      state_of_mind:  new UntypedFormControl(null, [Validators.required, Validators.maxLength(250)]),
+      conducts_and_non_verbal_languages:  new UntypedFormControl(null, [Validators.required, Validators.maxLength(250)]),
+      observations:  new UntypedFormControl(null, [Validators.required, Validators.maxLength(250)]),
     });
   }
 
