@@ -1,6 +1,6 @@
 //File generated by vaweei CLI
 import { Component, OnInit, Output, EventEmitter, ViewChild, ElementRef, HostListener, OnDestroy, Input, AfterViewInit } from '@angular/core';
-import { FormGroup, FormControl, Validators, FormBuilder } from '@angular/forms';
+import { UntypedFormGroup, UntypedFormControl, Validators, UntypedFormBuilder } from '@angular/forms';
 import { ActivatedRoute, Router } from '@angular/router'; 
 
 //SERVICES
@@ -19,7 +19,7 @@ import { filter } from 'rxjs/operators';
 export class TrackingUpdateFormViewComponent implements OnInit, OnDestroy, AfterViewInit {
 
   @ViewChild('firstInput', { static: false }) firstInput!: ElementRef;
-  formGroup: FormGroup;
+  formGroup: UntypedFormGroup;
   @Input() modalConfigParent: any;
   @Input() nameForm: String = '';
 
@@ -31,13 +31,13 @@ export class TrackingUpdateFormViewComponent implements OnInit, OnDestroy, After
     private utilService: UtilService,
     private router : Router,
     private route: ActivatedRoute,
-    private fb: FormBuilder,
+    private fb: UntypedFormBuilder,
   ) {
     this.setFocus();
     this.formGroup = this.fb.group({
-      id: new FormControl(null, [Validators.required, Validators.maxLength(250)]),
-      code: new FormControl(null, [Validators.required, Validators.maxLength(250)]),
-      name: new FormControl(null, [Validators.required, Validators.maxLength(250)]),
+      id: new UntypedFormControl(null, [Validators.required, Validators.maxLength(250)]),
+      code: new UntypedFormControl(null, [Validators.required, Validators.maxLength(250)]),
+      name: new UntypedFormControl(null, [Validators.required, Validators.maxLength(250)]),
     });
   }
 
