@@ -5,26 +5,18 @@ import { ReactiveFormsModule, FormsModule } from '@angular/forms';
 import { EmotionsFormViewComponent } from './emotions-form-view.component';
 
 @NgModule({
-
-  declarations: [
-    EmotionsFormViewComponent,
-  ],
-
-  entryComponents: [
-  ],
-
-  providers: [
-  ],
-
-  imports: [
-    ReactiveFormsModule,
-    FormsModule,
-    CommonModule,
-  ],
-
-  exports: [
-    EmotionsFormViewComponent
-  ]
+    declarations: [
+        EmotionsFormViewComponent,
+    ],
+    providers: [],
+    imports: [
+        ReactiveFormsModule,
+        FormsModule,
+        CommonModule,
+    ],
+    exports: [
+        EmotionsFormViewComponent
+    ]
 })
 
 export class EmotionsFormViewModule { }
