@@ -5,26 +5,18 @@ import { ReactiveFormsModule, FormsModule } from '@angular/forms';
 import { SchoolAreaFormViewComponent } from './school_area-form-view.component';
 
 @NgModule({
-
-  declarations: [
-    SchoolAreaFormViewComponent,
-  ],
-
-  entryComponents: [
-  ],
-
-  providers: [
-  ],
-
-  imports: [
-    ReactiveFormsModule,
-    FormsModule,
-    CommonModule,
-  ],
-
-  exports: [
-    SchoolAreaFormViewComponent
-  ]
+    declarations: [
+        SchoolAreaFormViewComponent,
+    ],
+    providers: [],
+    imports: [
+        ReactiveFormsModule,
+        FormsModule,
+        CommonModule,
+    ],
+    exports: [
+        SchoolAreaFormViewComponent
+    ]
 })
 
 export class SchoolAreaFormViewModule { }
