@@ -5,26 +5,18 @@ import { ReactiveFormsModule, FormsModule } from '@angular/forms';
 import { SocialAreaFormViewComponent } from './social_area-form-view.component';
 
 @NgModule({
-
-  declarations: [
-    SocialAreaFormViewComponent,
-  ],
-
-  entryComponents: [
-  ],
-
-  providers: [
-  ],
-
-  imports: [
-    ReactiveFormsModule,
-    FormsModule,
-    CommonModule,
-  ],
-
-  exports: [
-    SocialAreaFormViewComponent
-  ]
+    declarations: [
+        SocialAreaFormViewComponent,
+    ],
+    providers: [],
+    imports: [
+        ReactiveFormsModule,
+        FormsModule,
+        CommonModule,
+    ],
+    exports: [
+        SocialAreaFormViewComponent
+    ]
 })
 
 export class SocialAreaFormViewModule { }
