@@ -9,28 +9,20 @@ import { AssignTaskModalViewComponent } from './assign_task-modal-view.component
 import { AdvanceSearchModule } from 'src/app/utils/components/advance_search/advance_search.module';
 
 @NgModule({
-
-  declarations: [
-    AssignTaskModalViewComponent,
-  ],
-
-  entryComponents: [
-  ],
-
-  providers: [
-  ],
-
-  imports: [
-    ReactiveFormsModule,
-    FormsModule,
-    CommonModule,
-    NgSelectModule,
-    AdvanceSearchModule
-  ],
-
-  exports: [
-    AssignTaskModalViewComponent
-  ]
+    declarations: [
+        AssignTaskModalViewComponent,
+    ],
+    providers: [],
+    imports: [
+        ReactiveFormsModule,
+        FormsModule,
+        CommonModule,
+        NgSelectModule,
+        AdvanceSearchModule
+    ],
+    exports: [
+        AssignTaskModalViewComponent
+    ]
 })
 
 export class AssignTaskModalViewModule { }
