@@ -8,28 +8,20 @@ import { GenericModalComponent } from './generic-modal.component';
 
 import { TaskTemplateListViewModule } from 'src/app/components/psychotherapy/tasks/task_template-list-view/task_template-list-view.module';
 @NgModule({
-
-  declarations: [
-    GenericModalComponent,
-  ],
-
-  entryComponents: [
-  ],
-
-  providers: [
-  ],
-
-  imports: [
-    ReactiveFormsModule,
-    FormsModule,
-    CommonModule,
-    NgSelectModule,
-    TaskTemplateListViewModule
-  ],
-
-  exports: [
-    GenericModalComponent
-  ]
+    declarations: [
+        GenericModalComponent,
+    ],
+    providers: [],
+    imports: [
+        ReactiveFormsModule,
+        FormsModule,
+        CommonModule,
+        NgSelectModule,
+        TaskTemplateListViewModule
+    ],
+    exports: [
+        GenericModalComponent
+    ]
 })
 
 export class GenericModalModule { }
