@@ -10,30 +10,22 @@ import { MaterialAllModule } from 'src/material.module'
 import { AdvanceSearchModule } from 'src/app/utils/components/advance_search/advance_search.module';
 
 @NgModule({
-
-  declarations: [
-    HeaderComponent,
-  ],
-
-  entryComponents: [
-  ],
-
-  providers: [
-  ],
-
-  imports: [
-    ReactiveFormsModule,
-    FormsModule,
-    CommonModule,
-    MaterialAllModule,
-    ColorCircleModule,
-    RouterModule,
-    AdvanceSearchModule
-  ],
-
-  exports: [
-    HeaderComponent
-  ]
+    declarations: [
+        HeaderComponent,
+    ],
+    providers: [],
+    imports: [
+        ReactiveFormsModule,
+        FormsModule,
+        CommonModule,
+        MaterialAllModule,
+        ColorCircleModule,
+        RouterModule,
+        AdvanceSearchModule
+    ],
+    exports: [
+        HeaderComponent
+    ]
 })
 
 export class HeaderModule { }
